+++ src/main/java/solutions/fairdata/openrefine/metadata/commands/request/audit/AuditRequest.java
@@ -20,38 +20,19 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package solutions.fairdata.openrefine.metadata.model;
+package solutions.fairdata.openrefine.metadata.commands.request.audit;
 
-import com.google.refine.model.OverlayModel;
-import com.google.refine.model.Project;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import solutions.fairdata.openrefine.metadata.dto.audit.AuditLogDTO;
-import solutions.fairdata.openrefine.metadata.dto.config.ProjectConfigDTO;
+import solutions.fairdata.openrefine.metadata.dto.audit.EventType;
 
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @Setter
-public class MetadataOverlayModel implements OverlayModel {
-
-    private ProjectConfigDTO projectData;
-    private AuditLogDTO projectLog;
-
-    @Override
-    public void onBeforeSave(Project project) {
-        // noop
-    }
-
-    @Override
-    public void onAfterSave(Project project) {
-        // noop
-    }
-
-    @Override
-    public void dispose(Project project) {
-        // noop
-    }
+public class AuditRequest {
+    private EventType eventType;
+    private String message;
 }